Arena revenge system, real dungeon combat, alliance browse, + fixes
Arena:
- Fix "find opponent" button: answer callback immediately so the
  spinner never times out while matchmaking hits the DB.
- Attack logs now record attacker_label + attacker_power, and the
  arena home shows attacker name / power / gold looted per raid.
- Revenge: 3-day window on real losing raids. New "انتقام‌ها" panel
  lists open revenges with a countdown; taking one is atomic
  (mark_revenge_taken rolls back if the fight can't proceed).

Dungeon:
- Replace the bare daily dungeon with a real boss fight via
  resolve_duel — boss scaled to player power x per-dungeon multiplier,
  win/lose narrative, and a 30% consolation reward on a loss.

Alliance:
- Join flow is now a paginated browse list (by treasury) plus
  partial-name search, instead of an exact-name text prompt.
  Adds list_alliances_page / search_alliances / join_alliance_by_id.

Fixes / balance:
- Building collector no longer shows a fractional pending amount when
  the cap is binding (int-cast in pending_amount).
- Energy: MAX_ENERGY 20 -> 50, regen 12 -> 6 min (full in ~5h) for
  longer sessions; migration 0035 alters the default and adds the new
  AttackLog fields.

+++ game/buildings.py
@@ -141,7 +141,7 @@ def pending_amount(building: Building) -> int:
     rate = cfg["rate_per_hour"] * building.level * bonus
     cap = cfg["cap_base"] * building.level * bonus
     elapsed_hours = (timezone.now() - building.last_collected_at).total_seconds() / 3600
-    return min(cap, math.floor(rate * max(elapsed_hours, 0)))
+    return int(min(cap, math.floor(rate * max(elapsed_hours, 0))))
 
 
 def collect(user: User, building: Building) -> tuple[int, str]:
